+++ aito/cli/sub_commands/__init__.py
@@ -1,7 +1,0 @@
-from .convert_sub_command import ConvertSubCommand
-from .database_sub_command import ConfigureSubCommand, QuickAddTableSubCommand, CreateTableSubCommand, \
-    DeleteTableSubCommand, CopyTableSubCommand, RenameTableSubCommand, ShowTablesSubCommand, DeleteDatabaseSubCommand, \
-    UploadEntriesSubCommand, UploadBatchSubCommand, UploadFileSubCommand, UploadDataFromSQLSubCommand, \
-    QuickAddTableFromSQLSubCommand
-from .infer_table_schema_sub_command import InferTableSchemaSubCommand
-from .sub_command import SubCommand