+++ docs/source/changelog.rst
@@ -72,9 +72,11 @@ CLI
 - Added the following commands:
 
   - ``configure``: configure your Aito instance
+  - ``get-table``: return the schema of the specified table
   - ``show-tables``: show the existing tables in the Aito instance
   - ``copy-table``: copy a table
   - ``rename-table``: rename a table
+  - ``get-datbase``: return the schema of the database
 
 - Removed dotenv file support (***-e*** flag).